closes #188

+++ src/components/views/VotingResults.tsx
@@ -14,6 +14,7 @@ import { useNavigate, useParams } from "react-router-dom";
 import { Client } from "@stomp/stompjs";
 import { Chip, CircularProgress, Progress } from "@nextui-org/react";
 import { TailSpin, ThreeDots } from "react-loader-spinner";
+import ErrorMessageModal from "components/ui/ErrorMessageModal";
 
 const VotingResults = () => {
   const { gameId, setGameId } = useParams();
@@ -33,6 +34,10 @@ const VotingResults = () => {
   const [roundDurationSeconds, setRoundDurationSeconds] = useState(20);
   const [pageLoading, setPageLoading] = useState(true);
   let timerId;
+  const [errorModalOpen, setErrorModalOpen] = useState(false);
+  const [errorMessage, setErrorMessage] = useState("");
+  const [summaryId, setSummaryId] = useState(null);
+  let client = new Client();
 
   const openNotification = (message: string) => {
     toast.info(message, {autoClose: 3000});
@@ -161,7 +166,6 @@ const VotingResults = () => {
   }, []);
 
   useEffect(() => {
-    let client = new Client();
     const websocketUrl = getWebsocketDomain();
     client.configure({
       brokerURL: websocketUrl,
@@ -174,9 +178,19 @@ const VotingResults = () => {
             stopInactivityTimer();
             navigate(`/game/${gameId}/`);
           } else if (messageParsed.status === "game_over") {
-            stopInactivityTimer();
-            localStorage.clear();
-            navigate(`/gamesummary/${messageParsed.summaryId}/`);
+            if (currQuestNr !== totalQuests || remainingTime > 2) {
+              stopInactivityTimer();
+              client && client.deactivate();
+              localStorage.setItem("submissionDone", "false");
+              console.log("Game ended prematurely");
+              setErrorMessage("The game ended prematurely since less than three participants are remaining. You are being transferred to the summary page to see all completed rounds.");
+              setErrorModalOpen(true);
+              setSummaryId(messageParsed.summaryId);
+            } else {
+              stopInactivityTimer();
+              localStorage.clear();
+              navigate(`/gamesummary/${messageParsed.summaryId}/`);
+            }
           } else if (messageParsed.status === "left") {
             openNotification(messageParsed.username + " left");
           }
@@ -217,6 +231,13 @@ const VotingResults = () => {
 
     return (remainingTime / roundDurationSeconds) * 100;
   };
+
+  function handlePrematureGameEnd() {
+    setErrorModalOpen(false)
+    localStorage.clear();
+    navigate("/gamesummary/" + summaryId);
+  }
+
   const circularProgressStyles = {
     svg: "w-368 h-36 drop-shadow-md",
     indicator: "stroke-white",
@@ -227,6 +248,11 @@ const VotingResults = () => {
 
   return (
     <div className="relative min-h-screen w-screen flex flex-col items-center">
+      {errorModalOpen && <ErrorMessageModal isOpen={errorModalOpen} onClose={() => handlePrematureGameEnd()} errorMessage={errorMessage} />}
+      <ToastContainer
+        pauseOnFocusLoss={false}
+        pauseOnHover={false}
+      />
       {pageLoading ? (
         <div className="absolute top-0 left-0 w-full h-full flex items-center justify-center bg-gray-900 bg-opacity-50">
           <div className="flex flex-col items-center">
@@ -244,10 +270,6 @@ const VotingResults = () => {
       ) : (
         <BaseContainer
           size="large" className="flex flex-col items-center justify-center min-h-screen">
-          <ToastContainer
-            pauseOnFocusLoss={false}
-            pauseOnHover={false}
-          />
           <Progress
             aria-label="Progress"
             disableAnimation
